requirements.txt
tweepy==3.10.0
pyspark==2.4.7
kafka-python==2.0.2
python==3.7.5
findspark==1.4.2
pandas